Refactored the issues in correspondence with the introduction of portfolio directory
